use `compile_commands.json` for Cppcheck (#1716)
* suppress memleak warnings only for u_scanner.c

+++ src/r_draw.c
@@ -520,7 +520,7 @@ static void R_DrawFuzzColumn_orig(void)
       dest += linesize;             // killough 11/98
 
       // Clamp table lookup index.
-      fuzzpos &= (fuzzpos - FUZZTABLE) >> (8*sizeof fuzzpos-1); //killough 1/99
+      fuzzpos &= ((unsigned)(fuzzpos - FUZZTABLE)) >> (8*sizeof fuzzpos-1); //killough 1/99
     } 
   while (--count);
 
@@ -590,7 +590,7 @@ static void R_DrawFuzzColumn_block(void)
       }
 
       fuzzpos++;
-      fuzzpos &= (fuzzpos - FUZZTABLE) >> (8 * sizeof(fuzzpos) - 1);
+      fuzzpos &= ((unsigned)(fuzzpos - FUZZTABLE)) >> (8 * sizeof(fuzzpos) - 1);
     }
   while ((count -= ny) > 0);
